requirements.txt
uvicorn
Pillow
google-cloud-storage
fastai==2.8.5
fastapi==0.116.1
fastcore==1.8.15
fastdownload==0.0.7
fastjsonschema==2.21.1
fastprogress==1.0.3
fastrlock==0.8.3
fasttext==0.9.3
fasttransform==0.0.2
fastuuid==0.14.0
featuretools==1.31.0
ffmpy==0.6.1
filelock==3.20.0
filetype==1.2.0
jax==0.5.2
jaxlib==0.5.1
jeepney==0.9.0
jieba==0.42.1
Jinja2==3.1.6
jiter==0.10.0
jmespath==1.0.1
joblib==1.5.2
json5==0.12.1
jsonpatch==1.33
jsonpickle==4.1.1
jsonpointer==3.0.0
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
keras==3.8.0
keras-core==0.1.7
keras-cv==0.9.0
keras-hub==0.18.1
keras-nlp==0.18.1
keras-tuner==1.4.7
mypy_extensions==1.1.0
namex==0.1.0
narwhals==1.48.1
natsort==8.4.0
nbdev==2.4.6
nbformat==5.10.4
ndindex==1.10.0
nest-asyncio==1.6.0
networkx==3.5
nibabel==5.3.2
nilearn==0.10.4
ninja==1.13.0
numba==0.60.0
numexpr==2.11.0
numpy==1.26.4
oauth2client==4.1.3
oauthlib==3.3.1
odfpy==1.4.1
olefile==0.47
openpyxl==3.1.5
pandas==2.2.3
pandas-datareader==0.10.0
pandas-gbq==0.29.2
pandas-stubs==2.2.2.240909
pandasql==0.7.3
pandocfilters==1.5.1
pathspec==0.12.1
patsy==1.0.1
pdf2image==1.17.0
protobuf
python-lsp-jsonrpc==1.1.2
python-lsp-server==1.13.1
python-multipart==0.0.20
python-slugify==8.0.4
python-snappy==0.7.3
python-utils==3.9.1
pytokens==0.3.0
pytools==2025.2.5
regex==2025.11.3
requests==2.32.5
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
requirements-parser==0.9.0
safetensors==0.5.3
scikit-image==0.25.2
scikit-learn==1.2.2
scipy==1.15.3
sklearn-compat==0.1.3
sklearn-pandas==2.2.0
sympy==1.13.1
tensorflow==2.18.0
tensorflow-hub==0.16.1
tensorflow-io==0.37.1
tensorflow-metadata==1.17.2
tf-slim==1.1.0
tf_keras==2.18.0
transformers==4.53.3